initial keyboarding works
