+++ packages/web-components/fast-components/src/index.ts
@@ -8,6 +8,8 @@ export * from "./design-system-provider";
 export * from "./dialog";
 export * from "./divider";
 export * from "./flipper";
+export * from "./menu";
+export * from "./menu-item";
 export * from "./patterns";
 export * from "./progress";
 export * from "./radio";
