+++ packages/web-components/fast-components/src/menu-item/fixtures/menu-item.html
