+++ packages/web-components/fast-components/src/menu-item/menu-item.stories.ts
