+++ packages/web-components/fast-components/src/menu-item/menu-item.styles.ts
