+++ packages/web-components/fast-components/src/menu/menu.template.ts
@@ -5,9 +5,10 @@ export const MenuTemplate = html<Menu>`
 <template
     role="menu"
     @keydown=${(x, c) => x.handleMenuKeyDown(c.event as KeyboardEvent)}
->
+    @focusout=${(x, c) => x.handleFocusOut(c.event as FocusEvent)}
+    >
     <slot
         ${ref("items")}
     ></slot>
-</template
+</template>
 `;
